requirements_r32.txt
# requirements_r32_google.txt
requests>=2.31.0
pandas>=2.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
pyarrow>=13.0.0  # Optionnel, pour BigQuery